REFACTOR: move class LearnerInspector to new _learner_inspector.py
