+++ src/facet/inspection/__init__.py
@@ -7,3 +7,4 @@ of a learner pipeline which has been fitted using cross-validation.
 """
 from ._explainer import *
 from ._inspection import *
+from ._learner_inspector import *
